git: Fix issue with removed remote
In order to prevent commiting onto a worktree that was added via a
single commit, the remote url was being removed during the clone procedure.

However, this made it so we couldn't identify which url/upstream repository the
bare_git belonged to.

Since identification is crucial, the origin will no longer be removed.
Changes to repositories checked out in a detached commit will now fail.
This is now intended since in a git workflow, it doesn't make sense to
commit on a repository that has a detached head.

Signed-off-by: Bruno Mauricio <brunoasmauricio@gmail.com>