Gitall: Rework to fix "no commit" issue
"no commit" has been enforced by removing .git
However, this makes it impossible (or extremely hard) to know if a
folder is a git repository.

In order to overcome this, instead of gitall running on all git
repositories, it now only runs on the folders inside the project that we
know are in fact git repositories.

Git repositories that are not a dependency (aka not cloned by
ProjectBase) cannot be managed by ProjectBase

Signed-off-by: Bruno Mauricio <brunoasmauricio@gmail.com>