CMake: Macros to setup configure time files
Some test/executable files will not be generated by the build system,
but instead need to be copied over from the repository, to the
appropriate folder (executables or tests).

Macros were added to support this

It is a faster way to perform a common operation than adding the
commands to "after_build" in configuration.
Plus, the "after_build" should be used for system configuration or to
call external tools, not to add on top of the existing build system.

Signed-off-by: Bruno Mauricio <brunoasmauricio@gmail.com>